% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data74.R
\docType{data}
\name{condem74}
\alias{condem74}
\title{Hydrologically conditioned elevation for ONS station 74.}
\format{
A RasterLayer with elevation in m.
\describe{
\item{crs}{coordinate reference system is '+proj=longlat +datum=WGS84'}
\item{dimensions}{1957 rows, 3329 columns, 6514853 cells}
\item{resolution}{0.0008333333 arc degrees, ~90 m at equator}
}
}
\source{
\url{https://www.hydrosheds.org/downloads}
}
\usage{
condem74
}
\description{
As of Sep 2020.
}
\keyword{datasets}
