% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newave-data.R
\name{confhd_data}
\alias{confhd_data}
\title{Get data from \code{CONFHD.DAT} file for the a given year and month}
\source{
Data are available in \url{https://www.ccee.org.br/ccee/documentos}
since July/2012.
}
\usage{
confhd_data(
  YYYYMM = format.Date(Sys.Date() - 40, "\%Y\%m"),
  confhd_path = TRUE
)
}
\arguments{
\item{YYYYMM}{character vector with year (YYYY) and month (MM). Accepted
formats are for example "201809", "2018", "2018.02", "2018-03",
"2018-04-01", "2018/05".}

\item{confhd_path}{logical. Default: TRUE, will return the path
to \code{CONFHD.DAT} file, otherwise the path to the temporary directory of
extracted data.}
}
\value{
a \link[tibble:tibble-package]{tibble} with tidy data.
}
\description{
Get data from \code{CONFHD.DAT} file for the a given year and month
}
\note{
A data request for the current month may not yet be available for UHEs
configuration data. That is why it is recommended to use the date in the
previous 40 days to ensure that the data will be available.
}
\examples{
if(FALSE){
# Get the table for the previous month to ensure that the data already exists
uhes_info <- confhd_data(format.Date(Sys.Date()-40, "\%Y\%m"))
}

}
