% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station-polygon-operations.R
\name{elev_bands}
\alias{elev_bands}
\title{Fraction of precipitation and catchment area by elevation bands}
\usage{
elev_bands(con_dem, prec_raster = NULL, dz = 100, nbands = NULL)
}
\description{
Fraction of precipitation and catchment area by elevation bands
}
