% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HE-io.R
\name{extract_rar}
\alias{extract_rar}
\title{Extract files from a \code{.rar} file (only works on Linux)}
\usage{
extract_rar(rar_file, dest_dir = NULL, overwrite = FALSE, quiet = TRUE)
}
\arguments{
\item{rar_file}{a file path to a \code{file.rar}}

\item{dest_dir}{path to extract shape files}

\item{overwrite}{logical, use overwrite = TRUE to overwrite
pre-existent files.}

\item{quiet}{Hide printed output, messages, warnings, and errors
(TRUE, the default), or display them as they occur?}
}
\value{
character vector with shape files path
}
\description{
This function extract files from a \code{.rar} file
}
\details{
This function has the side effect of generating a
directory \code{BaciasHidrograficasONS_JUNTOS}' with the uncompressed files
in the \pkg{HEgis}
package directory (\code{system.file("extdata",package = "HEgis")}).
The files are extracted on \code{dest_dir} when it is not \code{NULL}.
}
\note{
This function is a wrapper to call \code{unrar}. It is required you
have \code{unrar} Linux library installed. You can install it with
'apt install unrar'.
}
\examples{
\dontrun{
if (interactive()) {
  bhs_rar_file <- system.file(
    "extdata",
    "BaciasHidrograficasONS_JUNTOS.rar",
    package = "HEgis"
  )
  shapes <- extract_rar(bhs_rar_file)
  shapes
}
}
}
