% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newave-data.R
\name{nw_down}
\alias{nw_down}
\title{Download zip file from link provided by \code{nw_urls()}}
\usage{
nw_down(link, confhd_path = TRUE)
}
\arguments{
\item{link}{url to download data}

\item{confhd_path}{logical. Default: TRUE, will return the path
to \code{CONFHD.DAT} file, otherwise the path to the temporary directory of
extracted data.}
}
\value{
Default is the path to \code{CONFHD.DAT} file, otherwise the
temporary directory of extracted data.
}
\description{
Download zip file from link provided by \code{nw_urls()}
}
