% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newave-data.R
\name{nw_urls}
\alias{nw_urls}
\title{Build the link to the monthly file of NEWAVE model CCEE deck of prices.}
\usage{
nw_urls(YYYYMM = "201809")
}
\arguments{
\item{YYYYMM}{character vector with year (YYYY) and month (MM). Accepted
formats are for example "201809", "2018", "2018.02", "2018-03",
"2018-04-01", "2018/05".}
}
\value{
A character string naming the URL of the zip file to be downloaded.
}
\description{
Build the link to the monthly file of NEWAVE model CCEE deck of prices.
}
\examples{
yyyymm = c("201809", "2018", "2018.02", "2018-03", "2018-04-01", "2018/05")
nw_urls(yyyymm)
}
