% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data74.R
\docType{data}
\name{poly74}
\alias{poly74}
\title{Catchment polygon of ONS station 74.}
\format{
A Simple Feature Polygon.
\describe{
\item{codONS}{Code ID from ONS station}
\item{codANA}{code ID by ANA}
\item{nome}{name}
\item{nomeOri}{name}
\item{adkm2}{drainage area in squared kilometers}
\item{volhm3}{reservoir volume in hm3}
\item{rio}{river name}
\item{cobacia}{code ID for the Ottobacia}
\item{tpopera}{operator}
}
}
\source{
The data were kindly provided by Saulo Aires
(Water Resources Specialist from ANA) through Prof. Carlos Lima
(UnB, e-mail in 2020-03-17).
}
\usage{
poly74
}
\description{
As of Sep 2020.
}
\note{
CRS is SIRGAS 2000 (EPSG: 4674)
}
\keyword{datasets}
