% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data74.R
\docType{data}
\name{precclim74}
\alias{precclim74}
\title{Annual climatology of Precipitation over catchment area of ONS station 74.}
\format{
A RasterLayer with total annual precipitation in mm.
\describe{
\item{crs}{coordinate reference system is '+proj=longlat +datum=WGS84'}
\item{dimensions}{14 rows, 19 columns, 266 cells}
\item{resolution}{0.25 arc degrees, ~25 km at equator}
}
}
\source{
\url{https://utexas.app.box.com/v/Xavier-etal-IJOC-DATA/}
}
\usage{
precclim74
}
\description{
As of Sep 2020.
}
\references{
https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/joc.4518
}
\keyword{datasets}
