% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-poly.R
\name{prep_poly_posto}
\alias{prep_poly_posto}
\title{Prepares the station's polygon to determine the spatial average of a
meterological variable.}
\usage{
prep_poly_posto(
  poly_station,
  dis.buf = 0.25,
  ref_crs = "+proj=longlat +datum=WGS84"
)
}
\arguments{
\item{poly_station}{a polygon of class \code{\link[sf]{sf}}}

\item{dis.buf}{numeric, default: 0.25 (degrees). When \code{dis.buf = 0},
the output polygon has only be converted the reprojected to \code{ref_crs}.}
}
\value{
object of class \code{\link[sf]{sf}}
}
\description{
Sets appropriate CRS to station's polygon and apply buffer.
}
