% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newave-data.R
\name{read_confhd}
\alias{read_confhd}
\title{Read data from text file \code{CONFHD.DAT}}
\usage{
read_confhd(confhd_file)
}
\arguments{
\item{confhd_file}{character, file path.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
Read data file that contains the names from Hydro Power plants and ID from
the streamflow gauge ('posto').
}
