% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station-polygon-operations.R
\name{z_bands}
\alias{z_bands}
\title{Fraction of the catchment covered by each Elevation band}
\usage{
z_bands(z, dz = 100, nbands = NULL)
}
\arguments{
\item{z}{raster or numeric vector}

\item{dz}{numeric scalar, interval (m) to elevation bands. Calculates basin
area distributions within 100 m elevation by default.}

\item{nbands}{numeric scalar. Default: NULL (use \code{dz} to build elevation
bands).}
}
\value{

}
\description{
Fraction of the catchment covered by each Elevation band
}
\examples{
\dontrun{
if(FALSE){
  poly_posto <- suppressMessages(extract_poly(station = 74))
   raster("~/Dropbox/datasets/GIS/hydrosheds/sa_con_3s_hydrosheds.grd"),
   poly_posto,
   dis.buf = 0
 )
  z_bands(z = con_posto, dz = 100)
  z_bands(z = con_posto, nbands = 5)
}
}
}
