% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newave-data.R
\name{confhd_data}
\alias{confhd_data}
\title{Get data from \code{CONFHD.DAT} file for the a given year and month}
\usage{
confhd_data(YYYYMM, confhd_path = TRUE)
}
\arguments{
\item{YYYYMM}{character vector with year (YYYY) and month (MM). Accepted
formats are for example "201809", "2018", "2018.02", "2018-03",
"2018-04-01", "2018/05".}

\item{confhd_path}{logical. Default: TRUE, will return the path
to \code{CONFHD.DAT} file, otherwise the path to the temporary directory of
extracted data.}
}
\value{

}
\description{
Get data from \code{CONFHD.DAT} file for the a given year and month
}
\examples{
uhes_info <- confhd_data(format.Date(Sys.Date(), "\%Y\%m"))
}
