% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HE-io.R
\name{extract_rar}
\alias{extract_rar}
\title{Extract files from a rar file (linux)}
\usage{
extract_rar(rar_file, dest_dir = NULL)
}
\arguments{
\item{rar_file}{a file path to a 'filename.rar'}

\item{dest_dir}{path to extract shape files
is both readable and writable. Default is FALSE.}
}
\value{
character vector with shape files path
}
\description{
This function extract files from a rar file
}
\details{
This function has the side effect of generating a
directory with the uncompressed files in the HEgis
package directory ('iextdata').
}
\examples{
\dontrun{
if (interactive()) {
  bhs_rar_file <- system.file(
    "extdata",
    "BaciasHidrograficasONS_JUNTOS.rar",
    package = "HEgis"
  )
  shapes <- extract_rar(bhs_rar_file)
  shapes
}
}
}
