% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEgis-package.R
\docType{package}
\name{HEgis}
\alias{HEgis}
\title{HEgis: Functions To Import GIS Data Needed For The HYDROENGIE Project}
\description{
The goal of HEgis is to prepare GIS data for use in HydroEngie R&D project.
}
\section{HEgis functions}{

The main functions are:
\itemize{
\item \code{\link{extract_rar}}
\item \code{\link{import_bhs_ons}}
}
}

