% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEgis-package.R
\docType{package}
\name{HEgis}
\alias{HEgis}
\alias{HEgis-package}
\title{HEgis: Functions To Import GIS Data Needed For The HYDROENGIE Project}
\description{
The goal of \pkg{HEgis} is to prepare GIS data for use in HydroEngie R&D project.
}
\section{HEgis functions}{


\itemize{
\code{\link{import_bhs_ons}}: a convenience function to import the
shapefile with basins of major hydroelectric power dams.
}
}

\author{
\strong{Maintainer}: Jônatan Tatsch \email{lhmet@ufsm.br}

}
