% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HE-io.R
\name{import_bhs_ons}
\alias{import_bhs_ons}
\title{Import shape file with watersheds polygons from ONS Hydro Power plants}
\source{
The data were kindly provided by Saulo Aires
(Water Resources Specialist from ANA) through Prof.
Carlos Lima (UnB, e-mail in 2020-03-17).
}
\usage{
import_bhs_ons(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical. Hide printed output data structure or display them as
they occur? Default FALSE.}
}
\value{
object of class \code{\link[sf]{sf}} when a layer was successfully
read; in case argument layer is missing and data source name (\code{dsn}) does not
contain a single layer, an object of class sf_layers is returned with the
layer names, each with their geometry type(s). Note that the number of
layers may also be zero.
}
\description{
Import shape file with watersheds polygons from ONS Hydro Power plants
}
\details{
This function look at the extdata folder of HEgis to extract the
file \code{BaciasHidrograficasONS_JUNTOS.rar} and import the shapefile of same
name in it.
}
\examples{
if (FALSE) {
    bhs_pols <- import_bhs_ons()
}
}
\author{
Jônatan Tatsch
}
