% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HE-io.R
\name{import_bhs_ons}
\alias{import_bhs_ons}
\title{Import shape files with watersheds polygons from ONS hydropower plants}
\source{
The data were kindly provided by Saulo Aires
(Water Resources Specialist from ANA) through Prof.
Carlos Lima (UnB, e-mail in 2020-03-17).
}
\usage{
import_bhs_ons(shape_file, verbose = FALSE)
}
\arguments{
\item{shape_file}{A character scalar. Path to shape file.}

\item{verbose}{A logical scalar. If TRUE prints data structure.}
}
\value{
tibble
}
\description{
Import shape files with watersheds polygons from ONS hydropower plants
}
\details{
This function was made to import a specific
file (e.g. 'BaciasHidrograficasONS_JUNTOS.shp'). It will be extended to
the files in 'BaciasHidrograficasONS_SEPARADO.rar'.
}
\examples{
\dontrun{
if (interactive()) {
  bhs_rar_file <- system.file(
    "extdata",
    "BaciasHidrograficasONS_JUNTOS.rar",
    package = "HEgis"
  )
  shapes <- extract_rar(bhs_rar_file)
  shapes
  bhs_shape_file <- shapes[grep("Bacias", fs::path_file(shapes))]
  pols_bhs <- import_bhs_ons(bhs_shape_file, verbose = TRUE)
}
}
}
\author{
Jonatan Tatsch
}
