% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-poly.R
\name{info_station}
\alias{info_station}
\title{Get basic info of ONS station}
\usage{
info_station(name_regex = "MUNHOZ")
}
\arguments{
\item{name_regex}{a string or regex to search for the name of the ONS station}
}
\value{
a \code{\link[tibble]{tibble}}
}
\description{
Get basic info of ONS station
}
\examples{
if(FALSE){
 info_station("MUNHOZ")
}
}
