% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station-polygon-operations.R
\name{prep_poly_posto}
\alias{prep_poly_posto}
\title{Prepares the station's polygon to determine the spatial average of a
meterological variable.}
\usage{
prep_poly_posto(
  poly_station,
  dis.buf = 0.25,
  ref_crs = "+proj=longlat +datum=WGS84"
)
}
\arguments{
\item{poly_station}{a polygon of class \code{\link[sf]{sf}}}

\item{dis.buf}{numeric, default: 0.25 (degrees). When \code{dis.buf = 0},
the only action over the \code{poly_station} is the conversion of coordinates to
\code{ref_crs}.}

\item{ref_crs}{character, the base coordinate reference system. Default:
'+proj=longlat +datum=WGS84'.}
}
\value{
object of class \code{\link[sf]{sf}}
}
\description{
Sets appropriate CRS to station's polygon and apply buffer.
}
